+++ packages/core/test/integrations/reactnativeinfo.test.ts
@@ -60,7 +60,7 @@ describe('React Native Info', () => {
         },
       },
       tags: {
-        hermes: 'true',
+        hermes: true,
       },
     });
   });
@@ -71,7 +71,7 @@ describe('React Native Info', () => {
     const actualEvent = await executeIntegrationFor({}, {});
 
     expectMocksToBeCalledOnce();
-    expect(actualEvent?.tags?.hermes).toEqual('true');
+    expect(actualEvent?.tags?.hermes).toBeTrue();
     expect(actualEvent?.contexts?.react_native_context).toEqual(
       expect.objectContaining({
         js_engine: 'hermes',