+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - Bump CLI from v2.46.0 to v2.47.0 ([#4979](https://github.com/getsentry/sentry-react-native/pull/4979))
   - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2470)
   - [diff](https://github.com/getsentry/sentry-cli/compare/2.46.0...2.47.0)
+- Bump Android SDK from v7.22.5 to v7.22.6 ([#4985](https://github.com/getsentry/sentry-react-native/pull/4985))
+  - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#7226)
+  - [diff](https://github.com/getsentry/sentry-java/compare/7.22.5...7.22.6)
 
 ## 6.16.1